Payment req modified

+++ src/main/java/com/example/demo/DTO/PaymentRequest.java
@@ -1,16 +1,14 @@
 package com.example.demo.DTO;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
+
 public class PaymentRequest {
     private String name;
     private String email;
-    private Double amount;
+    private long amount;
 }
